refactor: NodeIcon
